Add assertions
Use assert feature to document important assumptions that should hold in the code. For method in Parser, assert that the task type is correct before use. For first edited method in Ui, assert that array size is more than 0 before continuing with function. For second edited method in Ui, assert that task list should have a size more than zero before executing the code.

These assertions ensure that the programmer's assumptions are correct and stop the run if there are any discrepancies